refactor(ENTESB-12218): Refactor utility classes for OpenAPI 3.x support
