+++ app/server/api-generator/src/main/java/io/syndesis/server/api/generator/openapi/HttpStatus.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.syndesis.server.api.generator.swagger;
+package io.syndesis.server.api.generator.openapi;
 
 import java.util.Map;
 import java.util.stream.Collectors;
