+++ app/server/api-generator/src/main/java/io/syndesis/server/api/generator/openapi/OpenApiModelInfo.java
@@ -27,7 +27,7 @@ import io.apicurio.datamodels.openapi.v2.models.Oas20Document;
 import io.apicurio.datamodels.openapi.v3.models.Oas30Document;
 import io.syndesis.common.model.Violation;
 import io.syndesis.common.util.json.JsonUtils;
-import io.syndesis.server.api.generator.swagger.util.JsonSchemaHelper;
+import io.syndesis.server.api.generator.openapi.util.JsonSchemaHelper;
 import org.immutables.value.Value;
 
 /**
