+++ app/server/api-generator/src/test/java/io/syndesis/server/api/generator/openapi/util/XmlSchemaHelperTest.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.syndesis.server.api.generator.swagger.util;
+package io.syndesis.server.api.generator.openapi.util;
 
+import io.syndesis.server.api.generator.openapi.util.XmlSchemaHelper;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
