+++ app/server/api-generator/src/test/java/io/syndesis/server/api/generator/swagger/UnifiedXmlDataShapeGeneratorTest.java
@@ -28,7 +28,7 @@ import io.apicurio.datamodels.openapi.v2.models.Oas20Document;
 import io.apicurio.datamodels.openapi.v2.models.Oas20PathItem;
 import io.apicurio.datamodels.openapi.v2.models.Oas20Schema;
 import io.syndesis.common.util.json.JsonUtils;
-import io.syndesis.server.api.generator.swagger.util.XmlSchemaHelper;
+import io.syndesis.server.api.generator.openapi.util.XmlSchemaHelper;
 import org.dom4j.Document;
 import org.dom4j.DocumentHelper;
 import org.dom4j.Element;
